requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==9.0.0
spacy==3.2.0
SQLAlchemy==1.4.23
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
python-magic==0.4.24